Update set_face_normal function
