+++ includes/shape/hittable.h
@@ -6,7 +6,7 @@
 /*   By: jeongwpa <jeongwpa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 17:18:55 by jeongwpa          #+#    #+#             */
-/*   Updated: 2024/08/14 01:22:31 by jeongwpa         ###   ########.fr       */
+/*   Updated: 2024/08/14 17:49:56 by jeongwpa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,12 +64,14 @@ typedef struct s_hit_lst
 	int		capacity;
 }	t_hit_lst;
 
-void		set_face_normal( \
-	t_record *rec, t_ray const *r, t_vec3 outward_normal);
+
 t_hit_lst	*init_hittable_list(int capacity);
 void		add_hittable_list(t_hit_lst *list, t_hit *object);
 void		clear_hittable_list(t_hit_lst *list);
 t_bool		hit_shapes( \
 	t_hit_lst *list, t_ray const *ray, t_coll t, t_record *rec);
+
+void		set_face_normal( \
+	t_record *rec, t_ray const *r, t_vec3 outward_normal);
 
 #endif
